Move REST endpoint tests to explorer service
